nuevo

+++ singly_linked_lists/0-print_list.c
@@ -1,19 +1,22 @@
 #include "lists.h"
 #include <stdio.h>
 /**
- * print_list - function to print list
- * @h: pointer to the beginning of the list
- * Return: The count of elements
+ * print_list - function to print list all elements
+ * @h: pointer head 
+ * Return: number of nodes
  */
+
 size_t print_list(const list_t *h)
 {
-	size_t count = 0;
-		while (h != NULL)
+	int count = 0;
+
+	while (h != NULL)
 	{
 		if (h->str == NULL)
-		printf("[0] (nil)\n");
+			printf("[0] (nil)\n");
 		else
 			printf("[%u] %s\n", h->len, h->str);
+
 		h = h->next;
 		count++;
 	}
